requirements.txt
Django==3.1.3
Pillow==8.0.1
psycopg2-binary==2.8.6
django-oauth-toolkit==1.3.3
django-mptt==0.11.0
django-extensions==3.0.9
djangorestframework==3.12.2
django-minio-storage>=0.3.8